refactor(Moved various methods into the LegTransitionNotification class):
