+++ src/main/java/org/opentripplanner/middleware/models/LegTransitionNotification.java
@@ -1,9 +1,14 @@
 package org.opentripplanner.middleware.models;
 
 import org.opentripplanner.middleware.i18n.Message;
+import org.opentripplanner.middleware.otp.response.Leg;
 import org.opentripplanner.middleware.persistence.Persistence;
+import org.opentripplanner.middleware.tripmonitor.jobs.CheckMonitoredTrip;
 import org.opentripplanner.middleware.tripmonitor.jobs.NotificationType;
 import org.opentripplanner.middleware.triptracker.TravelerPosition;
+import org.opentripplanner.middleware.triptracker.TripStatus;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import javax.annotation.Nullable;
 import java.util.HashSet;
@@ -11,8 +16,17 @@ import java.util.Locale;
 import java.util.Set;
 
 import static com.mongodb.client.model.Filters.eq;
+import static org.opentripplanner.middleware.tripmonitor.jobs.NotificationType.ARRIVED_AND_MODE_CHANGE_NOTIFICATION;
+import static org.opentripplanner.middleware.tripmonitor.jobs.NotificationType.ARRIVED_NOTIFICATION;
+import static org.opentripplanner.middleware.tripmonitor.jobs.NotificationType.DEPARTED_NOTIFICATION;
+import static org.opentripplanner.middleware.triptracker.TravelerLocator.hasRequiredTransitLeg;
+import static org.opentripplanner.middleware.triptracker.TravelerLocator.hasRequiredTripStatus;
+import static org.opentripplanner.middleware.triptracker.TravelerLocator.hasRequiredWalkLeg;
+import static org.opentripplanner.middleware.triptracker.TravelerLocator.isApproachingEndOfLeg;
+import static org.opentripplanner.middleware.triptracker.TravelerLocator.isAtStartOfLeg;
 
 public class LegTransitionNotification {
+    private static final Logger LOG = LoggerFactory.getLogger(LegTransitionNotification.class);
 
     public String travelerName;
     public NotificationType notificationType;
@@ -86,5 +100,51 @@ public class LegTransitionNotification {
         });
 
         return notifyUsers;
+    }
+
+    /**
+     * If a traveler is on schedule and on either a walk or transit leg check for possible leg transition notification.
+     */
+    public static void checkForLegTransition(TripStatus tripStatus, TravelerPosition travelerPosition, MonitoredTrip trip) {
+        if (
+            hasRequiredTripStatus(tripStatus) &&
+            (hasRequiredWalkLeg(travelerPosition) || hasRequiredTransitLeg(travelerPosition))
+        ) {
+            NotificationType notificationType = getLegTransitionNotificationType(travelerPosition);
+            if (notificationType != null) {
+                try {
+                    new CheckMonitoredTrip(trip).processLegTransition(notificationType, travelerPosition);
+                } catch (CloneNotSupportedException e) {
+                    LOG.error("Error encountered while checking leg transition.", e);
+                }
+            }
+        }
+    }
+
+    /**
+     * Depending on the traveler's proximity to the start/end of a leg return the appropriate notification type.
+     */
+    private static NotificationType getLegTransitionNotificationType(TravelerPosition travelerPosition) {
+        if (isAtStartOfLeg(travelerPosition)) {
+            return DEPARTED_NOTIFICATION;
+        } else if (isApproachingEndOfLeg(travelerPosition)) {
+            if (hasModeChanged(travelerPosition)) {
+                return ARRIVED_AND_MODE_CHANGE_NOTIFICATION;
+            }
+            return ARRIVED_NOTIFICATION;
+        }
+        return null;
+    }
+
+    /**
+     * The traveler is at the end of the current leg and the mode has changed between this and the next leg.
+     */
+    private static boolean hasModeChanged(TravelerPosition travelerPosition) {
+        Leg nextLeg = travelerPosition.nextLeg;
+        Leg expectedLeg = travelerPosition.expectedLeg;
+        return
+            isApproachingEndOfLeg(travelerPosition) &&
+            nextLeg != null &&
+            !nextLeg.mode.equalsIgnoreCase(expectedLeg.mode);
     }
 }
